+++ contracts/accumulators/proto/curve/ManagedCurveLiquidityAccumulator.sol
@@ -32,22 +32,7 @@ contract ManagedCurveLiquidityAccumulator is AccessControlEnumerable, CurveLiqui
             maxUpdateDelay_
         )
         AccumulatorConfig(uint32(updateTheshold_), uint32(minUpdateDelay_), uint32(maxUpdateDelay_))
-    {
-        initializeRoles();
-    }
-
-    /**
-     * @notice Modifier to make a function callable only by a certain role. In
-     * addition to checking the sender's role, `address(0)` 's role is also
-     * considered. Granting a role to `address(0)` is equivalent to enabling
-     * this role for everyone.
-     */
-    modifier onlyRoleOrOpenRole(bytes32 role) {
-        if (!hasRole(role, address(0))) {
-            require(hasRole(role, msg.sender), "ManagedCurveLiquidityAccumulator: MISSING_ROLE");
-        }
-        _;
-    }
+    {}
 
     function canUpdate(bytes memory data) public view virtual override returns (bool) {
         // Return false if the message sender is missing the required role
@@ -78,26 +63,5 @@ contract ManagedCurveLiquidityAccumulator is AccessControlEnumerable, CurveLiqui
 
     function _updateThreshold() internal view virtual override returns (uint256) {
         return config.updateThreshold;
-    }
-
-    function initializeRoles() internal virtual {
-        // Setup admin role, setting msg.sender as admin
-        _setupRole(Roles.ADMIN, msg.sender);
-        _setRoleAdmin(Roles.ADMIN, Roles.ADMIN);
-
-        // CONFIG_ADMIN is managed by ADMIN
-        _setRoleAdmin(Roles.CONFIG_ADMIN, Roles.ADMIN);
-
-        // UPDATER_ADMIN is managed by ADMIN
-        _setRoleAdmin(Roles.UPDATER_ADMIN, Roles.ADMIN);
-
-        // ORACLE_UPDATER is managed by UPDATER_ADMIN
-        _setRoleAdmin(Roles.ORACLE_UPDATER, Roles.UPDATER_ADMIN);
-
-        // Hierarchy:
-        // ADMIN
-        //   - CONFIG_ADMIN
-        //   - UPDATER_ADMIN
-        //     - ORACLE_UPDATER
     }
 }