+++ contracts/accumulators/proto/offchain/ManagedOffchainPriceAccumulator.sol
@@ -18,9 +18,7 @@ contract ManagedOffchainPriceAccumulator is AccessControlEnumerable, OffchainPri
     )
         OffchainPriceAccumulator(averagingStrategy_, quoteToken_, updateTheshold_, minUpdateDelay_, maxUpdateDelay_)
         AccumulatorConfig(uint32(updateTheshold_), uint32(minUpdateDelay_), uint32(maxUpdateDelay_))
-    {
-        initializeRoles();
-    }
+    {}
 
     function canUpdate(bytes memory data) public view virtual override returns (bool) {
         // Return false if the message sender is missing the required role
@@ -50,26 +48,5 @@ contract ManagedOffchainPriceAccumulator is AccessControlEnumerable, OffchainPri
 
     function _updateThreshold() internal view virtual override returns (uint256) {
         return config.updateThreshold;
-    }
-
-    function initializeRoles() internal virtual {
-        // Setup admin role, setting msg.sender as admin
-        _setupRole(Roles.ADMIN, msg.sender);
-        _setRoleAdmin(Roles.ADMIN, Roles.ADMIN);
-
-        // CONFIG_ADMIN is managed by ADMIN
-        _setRoleAdmin(Roles.CONFIG_ADMIN, Roles.ADMIN);
-
-        // UPDATER_ADMIN is managed by ADMIN
-        _setRoleAdmin(Roles.UPDATER_ADMIN, Roles.ADMIN);
-
-        // ORACLE_UPDATER is managed by UPDATER_ADMIN
-        _setRoleAdmin(Roles.ORACLE_UPDATER, Roles.UPDATER_ADMIN);
-
-        // Hierarchy:
-        // ADMIN
-        //   - CONFIG_ADMIN
-        //   - UPDATER_ADMIN
-        //     - ORACLE_UPDATER
     }
 }